feat:심화 과제 컴포넌트 분리
