+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export * from './MonthView';
+export * from './WeekView';
